+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add support for all interfaces but `loopback` [#76](https://github.com/elad-bar/ha-edgeos/issues/76)
 - Improved WS connection management
+- Add 3 sensors for WS messages - Received, Ignored and Error
 
 ## 2.0.12